# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/vehicles": {
      "get": {
        "operationId": "VehiclesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Véhicules"
        ]
      },
      "post": {
        "operationId": "VehiclesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enrôlement véhicule + propriétaire (App Officier) ou enregistrement admin.",
        "tags": [
          "Véhicules"
        ]
      }
    },
    "/v1/vehicles/stats": {
      "get": {
        "operationId": "VehiclesController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Véhicules"
        ]
      }
    },
    "/v1/vehicles/{plate}": {
      "get": {
        "operationId": "VehiclesController_detail",
        "parameters": [
          {
            "name": "plate",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Recherche par plaque (scan OCR). 404 → proposer l'enrôlement.",
        "tags": [
          "Véhicules"
        ]
      }
    },
    "/v1/vehicles/{plate}/flag": {
      "post": {
        "operationId": "VehiclesController_flag",
        "parameters": [
          {
            "name": "plate",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Véhicules"
        ]
      }
    },
...
    "/v1/exports/contraventions.xlsx": {
      "get": {
        "operationId": "ExportsController_xlsx",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche : N°, plaque ou nom d'officier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statut",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "impayee",
                "partielle",
                "payee",
                "annulee"
              ]
            }
          },
          {
            "name": "zone_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "officer_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Exports"
        ]
      }
    },
    "/v1/exports/contraventions.pdf": {
      "get": {
        "operationId": "ExportsController_pdf",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche : N°, plaque ou nom d'officier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statut",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "impayee",
                "partielle",
                "payee",
                "annulee"
              ]
            }
          },
          {
            "name": "zone_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "officer_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Exports"
        ]
      }
    },
    "/v1/exports/finance.pdf": {
      "get": {
        "operationId": "ExportsController_financePdf",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Exports"
        ]
      }
    },
...
    "/v1/me/vehicles/{plate}": {
      "get": {
        "operationId": "UserAppController_myVehicle",
        "parameters": [
          {
            "name": "plate",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Espace Usager"
        ]
      }
    },
...
  },
  "info": {
    "title": "SEN Contraventions API",
    "description": "Backend unique de l’Admin Web, de l’App Officier et de l’Espace Usager.\n\n**Authentification** : `Authorization: Bearer <accessToken>` (15 min), renouvelé via `POST /v1/auth/refresh`.\n**Officier suspendu** : toute requête renvoie HTTP 423 `OFFICER_SUSPENDED` → verrouiller le terminal.\n**Temps réel** : Socket.IO, namespace `/ws`, `auth: { token }`. Événements : `contravention.created`, \n`payment.confirmed`, `infraction_types.updated`, `vehicle.flagged`, `notification.created`, `officer.suspended`.\n**Hors-ligne** : envoyer `clientUuid` (UUID v4) dans `POST /v1/contraventions` pour une synchronisation idempotente.\n**Montants** : entiers en FCFA. **Dates** : ISO 8601.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ContraventionItemDto": {
        "type": "object",
        "properties": {
          "infractionTypeId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "libelle": {
            "type": "string"
          },
          "icone": {
            "type": "string"
          },
          "montant": {
            "type": "number"
          }
        },
        "required": [
          "infractionTypeId",
          "code",
          "libelle",
          "icone",
          "montant"
        ]
      },
      "ContraventionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "numero": {
            "type": "string"
          },
          "dateHeure": {
            "format": "date-time",
            "type": "string"
          },
          "plaque": {
            "type": "string"
          },
          "vehicule": {
            "type": "string"
          },
          "proprietaire": {
            "type": "string"
          },
          "infraction": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContraventionItemDto"
            }
          },
          "officierId": {
            "type": "string",
            "nullable": true
          },
          "officier": {
            "type": "string",
            "nullable": true
          },
          "zoneId": {
            "type": "string"
          },
          "zone": {
            "type": "string"
          },
          "montantTotal": {
            "type": "number"
          },
          "montantPaye": {
            "type": "number"
          },
          "resteAPayer": {
            "type": "number"
          },
          "statut": {
            "enum": [
              "impayee",
              "partielle",
              "payee",
              "annulee"
            ],
            "type": "string"
          },
          "canal": {
            "enum": [
              "admin",
              "mobile"
            ],
            "type": "string"
          },
          "lieuTexte": {
            "type": "string",
            "nullable": true
          },
          "lat": {
            "type": "number",
            "nullable": true
          },
          "lng": {
            "type": "number",
            "nullable": true
          },
          "photoPreuveUrl": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "qrPayload": {
            "type": "string",
            "description": "Contenu à encoder dans le QR code de l'avis"
          }
        },
        "required": [
          "id",
          "numero",
          "dateHeure",
          "plaque",
          "vehicule",
          "proprietaire",
          "infraction",
          "items",
          "officierId",
          "officier",
          "zoneId",
          "zone",
          "montantTotal",
          "montantPaye",
          "resteAPayer",
          "statut",
          "canal",
          "lieuTexte",
          "lat",
          "lng",
          "photoPreuveUrl",
          "notes",
          "qrPayload"
        ]
      },
...
      "VehicleDocumentDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "ASS",
              "VT",
              "CG"
            ]
          },
          "label": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "expiration": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "code",
          "label",
          "ok",
          "expiration"
        ]
      },
      "OwnerSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "cni": {
            "type": "string"
          },
          "telephone": {
            "type": "string"
          },
          "quartier": {
            "type": "string",
            "nullable": true
          },
          "hasAppAccount": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "prenom",
          "nom",
          "cni",
          "telephone",
          "quartier",
          "hasAppAccount"
        ]
      },
      "VehicleDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "plaque": {
            "type": "string"
          },
          "marque": {
            "type": "string"
          },
          "modele": {
            "type": "string"
          },
          "couleur": {
            "type": "string",
            "nullable": true
          },
          "annee": {
            "type": "number",
            "nullable": true
          },
          "type": {
            "enum": [
              "VP",
              "Moto",
              "Camion",
              "Bus"
            ],
            "type": "string"
          },
          "carburant": {
            "type": "string",
            "nullable": true
          },
          "chassis": {
            "type": "string",
            "nullable": true
          },
          "flagged": {
            "type": "boolean"
          },
          "flagReason": {
            "type": "string",
            "nullable": true
          },
          "alerte": {
            "type": "boolean",
            "description": "Document expiré, amende impayée ou véhicule signalé"
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VehicleDocumentDto"
            }
          },
          "proprietaire": {
            "$ref": "#/components/schemas/OwnerSummaryDto"
          },
          "totalDu": {
            "type": "number",
            "description": "Somme restant due sur les amendes impayées / partielles"
          },
          "impayees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContraventionDto"
            }
          },
          "historique": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContraventionDto"
            }
          }
        },
        "required": [
          "id",
          "plaque",
          "marque",
          "modele",
          "couleur",
          "annee",
          "type",
          "carburant",
          "chassis",
          "flagged",
          "flagReason",
          "alerte",
          "documents",
          "proprietaire",
          "totalDu",
          "impayees",
          "historique"
        ]
      },
      "OwnerInputDto": {
        "type": "object",
        "properties": {
          "prenom": {
            "type": "string"
          },
          "nom": {
            "type": "string"
          },
          "cni": {
            "type": "string",
            "description": "Si un propriétaire avec cette CNI existe déjà, il est réutilisé."
          },
          "telephone": {
            "type": "string"
          },
          "adresse": {
            "type": "string"
          },
          "quartier": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "photoCniUrl": {
            "type": "string",
            "description": "URL de la photo de la CNI (POST /uploads)"
          }
        },
        "required": [
          "prenom",
          "nom",
          "cni",
          "telephone"
        ]
      },
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "plaque": {
            "type": "string"
          },
          "marque": {
            "type": "string"
          },
          "modele": {
            "type": "string"
          },
          "couleur": {
            "type": "string"
          },
          "annee": {
            "type": "number"
          },
          "type": {
            "enum": [
              "VP",
              "Moto",
              "Camion",
              "Bus"
            ],
            "type": "string"
          },
          "carburant": {
            "type": "string"
          },
          "chassis": {
            "type": "string"
          },
          "assuranceExp": {
            "type": "string"
          },
          "visiteTechExp": {
            "type": "string"
          },
          "carteGriseOk": {
            "type": "boolean"
          },
          "ownerId": {
            "type": "string",
            "description": "Propriétaire existant…"
          },
          "owner": {
            "description": "…ou nouveau propriétaire (enrôlement terrain, étape 2)",
            "allOf": [
              {
                "$ref": "#/components/schemas/OwnerInputDto"
              }
            ]
          }
        },
        "required": [
          "plaque",
          "marque",
          "modele",
          "type"
        ]
      },
      "FlagVehicleDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  }
}